# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AdSetResponse:
    properties:
      attribution_spec: {}
      bid_strategy:
        example: LOWEST_COST_WITHOUT_CAP
        type: string
      budget_remaining:
        example: "50000"
        type: string
      campaign_id:
        example: "2381234567890"
        type: string
      created_time:
        example: "2026-05-11T10:00:00Z"
        type: string
      daily_budget:
        example: "50000"
        type: string
      effective_status:
        example: ACTIVE
        type: string
      end_time:
        example: "2026-06-11T10:00:00Z"
        type: string
      id:
        example: "2389876543210"
        type: string
      lifetime_budget:
        example: "0"
        type: string
      name:
        example: AdSet Leads USA
        type: string
      start_time:
        example: "2026-05-11T10:00:00Z"
        type: string
      status:
        example: ACTIVE
        type: string
      updated_time:
        example: "2026-05-11T12:00:00Z"
        type: string
    type: object
...
  dto.RoleBrief:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.UserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      role_ids:
        items:
          type: integer
        type: array
    required:
    - email
    - name
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/dto.RoleBrief'
        type: array
    type: object
...
  response.ErrorResponse:
    properties:
      errors: {}
      message:
        example: Error message
        type: string
      success:
        example: false
        type: boolean
    type: object
  response.Meta:
    properties:
      last_page:
        type: integer
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  response.MetaPaging:
    properties:
      cursors:
        properties:
          after:
            type: string
          before:
            type: string
        type: object
      has_next:
        type: boolean
      has_previous:
        type: boolean
    type: object
  response.PaginationMeta:
    properties:
      last_page:
        type: integer
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  response.PaginationResponse:
    properties:
      data: {}
      message:
        example: Success
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        example: true
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      paging:
        $ref: '#/definitions/response.MetaPaging'
      success:
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8888
info:
  contact: {}
  description: ERP Ready Backend for Meta Ads & Business Management
  title: Ads Backend API
  version: "1.0"
paths:
...
  /meta/adsets:
    get:
      consumes:
      - application/json
      description: Retrieve adsets from local database (synced from Meta)
      parameters:
      - description: Filter by campaign ID
        in: query
        name: campaign_id
        type: string
      - description: Filter by status (ACTIVE, PAUSED, etc)
        in: query
        name: status
        type: string
      - description: Search by adset name
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.AdSetResponse'
                  type: array
                meta:
                  $ref: '#/definitions/response.PaginationMeta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get AdSets
      tags:
      - Meta AdSets
...
  /meta/dashboard/ads:
    get:
      consumes:
      - application/json
      description: Returns ads with performance metrics joined from insights
      parameters:
      - description: Ad Account ID
        in: query
        name: ad_account_id
        type: string
      - description: Brand ID
        in: query
        name: brand_id
        type: integer
      - description: Filter by campaign ID
        in: query
        name: campaign_id
        type: string
      - description: Filter by ad set ID
        in: query
        name: adset_id
        type: string
      - description: Filter by ad status
        in: query
        name: status
        type: string
      - description: Search by ad name
        in: query
        name: search
        type: string
      - description: Filter by date start (YYYY-MM-DD)
        in: query
        name: date_start
        type: string
      - description: Filter by date stop (YYYY-MM-DD)
        in: query
        name: date_stop
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Ad Dashboard
      tags:
      - Meta Dashboard
  /meta/dashboard/adsets:
    get:
      consumes:
      - application/json
      description: Returns ad sets with performance metrics joined from insights
      parameters:
      - description: Ad Account ID
        in: query
        name: ad_account_id
        type: string
      - description: Brand ID
        in: query
        name: brand_id
        type: integer
      - description: Filter by campaign ID
        in: query
        name: campaign_id
        type: string
      - description: Filter by ad set status
        in: query
        name: status
        type: string
      - description: Search by ad set name
        in: query
        name: search
        type: string
      - description: Filter by date start (YYYY-MM-DD)
        in: query
        name: date_start
        type: string
      - description: Filter by date stop (YYYY-MM-DD)
        in: query
        name: date_stop
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Ad Set Dashboard
      tags:
      - Meta Dashboard
  /meta/dashboard/campaigns:
    get:
      consumes:
      - application/json
      description: Returns campaigns with performance metrics (spend, impressions,
        reach, actions) joined from insights and adsets
      parameters:
      - description: Ad Account ID
        in: query
        name: ad_account_id
        type: string
      - description: Brand ID
        in: query
        name: brand_id
        type: integer
      - description: Filter by campaign status (ACTIVE, PAUSED, etc)
        in: query
        name: status
        type: string
      - description: Search by campaign name
        in: query
        name: search
        type: string
      - description: Filter by date start (YYYY-MM-DD)
        in: query
        name: date_start
        type: string
      - description: Filter by date stop (YYYY-MM-DD)
        in: query
        name: date_stop
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Campaign Dashboard
      tags:
      - Meta Dashboard
...
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve all users
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by email
        in: query
        name: email
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginationResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UserResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get All Users
      tags:
      - User Management
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: Create User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Register User
      tags:
      - User Management
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - User Management
    get:
      consumes:
      - application/json
      description: Retrieve a user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get User by ID
      tags:
      - User Management
    put:
      consumes:
      - application/json
      description: Update a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Update User
      tags:
      - User Management
securityDefinitions:
  BearerAuth:
    description: Tuliskan 'Bearer ' diikuti dengan token JWT Anda.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"